Remaining libraries ported to AArch64
USB, Input, FS, FAT FS, Scheduler, Net and BT

+++ include/circle/input/linediscipline.h
@@ -2,7 +2,7 @@
 // linediscipline.h
 //
 // Circle - A C++ bare metal environment for Raspberry Pi
-// Copyright (C) 2017  R. Stange <rsta2@o2online.de>
+// Copyright (C) 2017-2018  R. Stange <rsta2@o2online.de>
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -39,7 +39,7 @@ public:
 	CLineDiscipline (CDevice *pInputDevice, CDevice *pOutputDevice);
 	~CLineDiscipline (void);
 
-	int Read (void *pBuffer, unsigned nCount);
+	int Read (void *pBuffer, size_t nCount);
 
 	void SetOptionRawMode (boolean bEnable);
 	void SetOptionEcho (boolean bEnable);
